refactor(substrate-core): split control.rs tests into separate module (issue 426)
